CEDS-1400 Rename and move files and method names
