+++ app/uk/gov/hmrc/exports/movements/controllers/MovementsSubmissionController.scala
@@ -21,11 +21,9 @@ import play.api.Logger
 import play.api.mvc._
 import uk.gov.hmrc.auth.core.AuthConnector
 import uk.gov.hmrc.exports.movements.config.AppConfig
-import uk.gov.hmrc.exports.movements.models.{
-  AuthorizedSubmissionRequest,
-  ErrorResponse,
-  ValidatedHeadersMovementsRequest
-}
+import uk.gov.hmrc.exports.movements.controllers.actions.Authenticator
+import uk.gov.hmrc.exports.movements.controllers.util.HeaderValidator
+import uk.gov.hmrc.exports.movements.models.{AuthorizedSubmissionRequest, ErrorResponse, ValidatedHeadersMovementsRequest}
 import uk.gov.hmrc.exports.movements.services.MovementsService
 import uk.gov.hmrc.http.HeaderCarrier
 
@@ -40,7 +38,7 @@ class MovementsSubmissionController @Inject()(
   headerValidator: HeaderValidator,
   movementsService: MovementsService,
   cc: ControllerComponents
-) extends ExportController(authConnector, cc) {
+) extends Authenticator(authConnector, cc) {
 
   def submitMovement(): Action[AnyContent] =
     authorisedAction(bodyParser = xmlOrEmptyBody) { implicit request =>
