+++ conf/app.routes
@@ -1,6 +1,6 @@
 # microservice specific routes
 
-POST        /customs-declare-exports/notifyMovement        uk.gov.hmrc.exports.movements.controllers.NotificationsController.saveMovement()
+POST        /customs-declare-exports/notifyMovement        uk.gov.hmrc.exports.movements.controllers.NotificationsController.saveNotification()
 
 POST        /save-movement-submission                      uk.gov.hmrc.exports.movements.controllers.MovementsSubmissionController.submitMovement()
