+++ test/component/uk/gov/hmrc/exports/movements/base/ComponentTestSpec.scala
@@ -28,7 +28,7 @@ import play.api.Application
 import play.api.inject.bind
 import play.api.inject.guice.GuiceApplicationBuilder
 import uk.gov.hmrc.exports.movements.models.MovementSubmissions
-import uk.gov.hmrc.exports.movements.repositories.{MovementNotificationsRepository, MovementsRepository}
+import uk.gov.hmrc.exports.movements.repositories.{MovementsRepository, NotificationsRepository}
 import utils.ExternalServicesConfig.{Host, Port}
 import utils.stubs.CustomsMovementsAPIService
 import utils.{AuthService, CustomsMovementsAPIConfig}
@@ -39,7 +39,7 @@ trait ComponentTestSpec
     extends FeatureSpec with GivenWhenThen with GuiceOneAppPerSuite with BeforeAndAfterAll with BeforeAndAfterEach
     with Eventually with MockitoSugar with Matchers with OptionValues with AuthService with CustomsMovementsAPIService {
 
-  private val mockMovementNotificationsRepository = mock[MovementNotificationsRepository]
+  private val mockMovementNotificationsRepository = mock[NotificationsRepository]
   private val mockMovementSubmissionsRepository = mock[MovementsRepository]
 
   override protected def beforeAll() {
@@ -75,7 +75,7 @@ trait ComponentTestSpec
   override implicit lazy val app: Application =
     GuiceApplicationBuilder()
       .overrides(bind[MovementsRepository].toInstance(mockMovementSubmissionsRepository))
-      .overrides(bind[MovementNotificationsRepository].toInstance(mockMovementNotificationsRepository))
+      .overrides(bind[NotificationsRepository].toInstance(mockMovementNotificationsRepository))
       .configure(
         Map(
           "microservice.services.auth.host" -> Host,
