+++ test/unit/uk/gov/hmrc/exports/movements/base/CustomsExportsBaseSpec.scala
@@ -29,7 +29,7 @@ import org.scalatestplus.play.guice.GuiceOneAppPerSuite
 import play.api.Application
 import play.api.i18n.MessagesApi
 import play.api.inject.guice.GuiceApplicationBuilder
-import play.api.inject.{bind, Injector}
+import play.api.inject.{Injector, bind}
 import play.api.libs.ws.WSClient
 import play.filters.csrf.{CSRFConfig, CSRFConfigProvider, CSRFFilter}
 import uk.gov.hmrc.auth.core._
@@ -37,7 +37,7 @@ import uk.gov.hmrc.exports.movements.config.AppConfig
 import uk.gov.hmrc.exports.movements.connectors.CustomsInventoryLinkingExportsConnector
 import uk.gov.hmrc.exports.movements.metrics.ExportsMetrics
 import uk.gov.hmrc.exports.movements.models.{CustomsInventoryLinkingResponse, MovementSubmissions}
-import uk.gov.hmrc.exports.movements.repositories.{MovementNotificationsRepository, MovementsRepository}
+import uk.gov.hmrc.exports.movements.repositories.{MovementsRepository, NotificationsRepository}
 
 import scala.concurrent.duration._
 import scala.concurrent.{ExecutionContext, Future}
@@ -48,13 +48,13 @@ trait CustomsExportsBaseSpec
     GuiceApplicationBuilder()
       .overrides(
         bind[AuthConnector].to(mockAuthConnector),
-        bind[MovementNotificationsRepository].to(mockMovementNotificationsRepository),
+        bind[NotificationsRepository].to(mockMovementNotificationsRepository),
         bind[MovementsRepository].to(mockMovementsRepository),
         bind[CustomsInventoryLinkingExportsConnector].to(mockCustomsInventoryLinkingConnector),
         bind[ExportsMetrics].to(mockMetrics)
       )
       .build()
-  val mockMovementNotificationsRepository: MovementNotificationsRepository = mock[MovementNotificationsRepository]
+  val mockMovementNotificationsRepository: NotificationsRepository = mock[NotificationsRepository]
   val mockMovementsRepository: MovementsRepository = mock[MovementsRepository]
   val mockCustomsInventoryLinkingConnector: CustomsInventoryLinkingExportsConnector =
     mock[CustomsInventoryLinkingExportsConnector]
