+++ test/utils/stubs/CustomsMovementsAPIService.scala
@@ -26,7 +26,7 @@ import integration.uk.gov.hmrc.exports.movements.base.WireMockRunner
 import play.api.http.ContentTypes
 import play.api.mvc.Codec
 import play.api.test.Helpers.{ACCEPTED, CONTENT_TYPE}
-import uk.gov.hmrc.exports.movements.controllers.CustomsHeaderNames
+import uk.gov.hmrc.exports.movements.controllers.util.CustomsHeaderNames
 import utils.{CustomsMovementsAPIConfig, MovementsTestData}
 
 trait CustomsMovementsAPIService extends WireMockRunner with MovementsTestData {
